+++ web/src/auth/SignInSignUpCommon.tsx
@@ -1,6 +1,6 @@
 import * as H from 'history'
 import * as React from 'react'
-import { VALID_USERNAME_REGEXP } from '../user'
+import { USERNAME_MAX_LENGTH, VALID_USERNAME_REGEXP } from '../user'
 
 export const PasswordInput: React.SFC<
     React.InputHTMLAttributes<HTMLInputElement> & { inputRef?: React.Ref<HTMLInputElement> }
@@ -37,6 +37,7 @@ export const UsernameInput: React.SFC<React.InputHTMLAttributes<HTMLInputElement
         placeholder="Username"
         spellCheck={false}
         pattern={VALID_USERNAME_REGEXP.toString().slice(1, -1)}
+        maxLength={USERNAME_MAX_LENGTH}
         autoComplete="username"
     />
 )